aleph/requirements.txt
asgiref==3.8.1
blessed==1.20.0
boto3==1.34.72
botocore==1.34.72
bpython==0.24
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==36.0.0
curtsies==0.4.2
cwcwidth==0.1.9
Django==5.0.3
django-filter==24.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
# dotenv==0.0.5
dynaconf==3.2.5
greenlet==3.0.3
hvac==2.1.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyocr==0.8.5
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyxdg==0.28
requests==2.31.0
s3transfer==0.10.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tesseract==0.1.3
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.24.0
wcwidth==0.2.13
